Improve dashboard layout and split test file
- Make all dashboard widgets full-width with uniform padding
- Expand heatmap to 31 weeks filling full content width
- Expand sparkline with proportional character widths and month labels
- Add two-column stats widget with new metrics: daily avg, sessions
  this week, total hours, top project, best day
- Align time summary deltas with fixed arrow column
- Split monolith test file into 9 focused files by domain
- Add 10 DashboardService tests

Closes #69
