+++ Tests/RockyCoreTests/SessionEditTests.swift
@@ -1,0 +1,293 @@
+import Testing
+import Foundation
+@testable import RockyCore
+
+@Suite("SessionService.editSession")
+struct SessionEditTests {
+    private func makeServices() -> (MockProjectRepository, MockSessionRepository, SessionService) {
+        let projectRepo = MockProjectRepository()
+        let sessionRepo = MockSessionRepository(projectRepository: projectRepo)
+        let sessionService = SessionService(repository: sessionRepo)
+        return (projectRepo, sessionRepo, sessionService)
+    }
+
+    private let cal = Calendar.current
+
+    private func insertSession(
+        _ sessionRepo: MockSessionRepository,
+        projectId: Int,
+        startHour: Int, startDay: Int = 6,
+        endHour: Int, endDay: Int = 6
+    ) async throws {
+        let start = cal.date(from: DateComponents(year: 2026, month: 3, day: startDay, hour: startHour))!
+        let end = cal.date(from: DateComponents(year: 2026, month: 3, day: endDay, hour: endHour))!
+        try await sessionRepo.insert(projectId: projectId, startTime: start, endTime: end)
+    }
+
+    @Test("edit with --start only updates start, keeps stop")
+    func editStartOnly() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await insertSession(sessionRepo, projectId: project.id, startHour: 10, endHour: 12)
+
+        let all = try await sessionRepo.getSessions(
+            from: cal.date(from: DateComponents(year: 2026, month: 3, day: 6))!,
+            to: cal.date(from: DateComponents(year: 2026, month: 3, day: 7))!
+        )
+        let sessionId = all[0].0.id
+        let originalEnd = all[0].0.endTime!
+
+        let newStart = cal.date(from: DateComponents(year: 2026, month: 3, day: 6, hour: 9))!
+        let updated = try await service.editSession(id: sessionId, newStart: newStart, newStop: nil, newDuration: nil)
+
+        #expect(updated.startTime == newStart)
+        #expect(updated.endTime == originalEnd)
+    }
+
+    @Test("edit with --stop only updates stop, keeps start")
+    func editStopOnly() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await insertSession(sessionRepo, projectId: project.id, startHour: 10, endHour: 12)
+
+        let all = try await sessionRepo.getSessions(
+            from: cal.date(from: DateComponents(year: 2026, month: 3, day: 6))!,
+            to: cal.date(from: DateComponents(year: 2026, month: 3, day: 7))!
+        )
+        let sessionId = all[0].0.id
+        let originalStart = all[0].0.startTime
+
+        let newStop = cal.date(from: DateComponents(year: 2026, month: 3, day: 6, hour: 14))!
+        let updated = try await service.editSession(id: sessionId, newStart: nil, newStop: newStop, newDuration: nil)
+
+        #expect(updated.startTime == originalStart)
+        #expect(updated.endTime == newStop)
+    }
+
+    @Test("edit with --start + --stop updates both")
+    func editStartAndStop() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await insertSession(sessionRepo, projectId: project.id, startHour: 10, endHour: 12)
+
+        let all = try await sessionRepo.getSessions(
+            from: cal.date(from: DateComponents(year: 2026, month: 3, day: 6))!,
+            to: cal.date(from: DateComponents(year: 2026, month: 3, day: 7))!
+        )
+        let sessionId = all[0].0.id
+
+        let newStart = cal.date(from: DateComponents(year: 2026, month: 3, day: 6, hour: 9))!
+        let newStop = cal.date(from: DateComponents(year: 2026, month: 3, day: 6, hour: 11))!
+        let updated = try await service.editSession(id: sessionId, newStart: newStart, newStop: newStop, newDuration: nil)
+
+        #expect(updated.startTime == newStart)
+        #expect(updated.endTime == newStop)
+    }
+
+    @Test("edit with --duration only keeps start, computes stop")
+    func editDurationOnly() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await insertSession(sessionRepo, projectId: project.id, startHour: 10, endHour: 12)
+
+        let all = try await sessionRepo.getSessions(
+            from: cal.date(from: DateComponents(year: 2026, month: 3, day: 6))!,
+            to: cal.date(from: DateComponents(year: 2026, month: 3, day: 7))!
+        )
+        let sessionId = all[0].0.id
+        let originalStart = all[0].0.startTime
+
+        let updated = try await service.editSession(id: sessionId, newStart: nil, newStop: nil, newDuration: 3600)
+
+        #expect(updated.startTime == originalStart)
+        #expect(abs(updated.endTime!.timeIntervalSince(originalStart) - 3600) < 1)
+    }
+
+    @Test("edit with --start + --duration sets start and computes stop")
+    func editStartAndDuration() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await insertSession(sessionRepo, projectId: project.id, startHour: 10, endHour: 12)
+
+        let all = try await sessionRepo.getSessions(
+            from: cal.date(from: DateComponents(year: 2026, month: 3, day: 6))!,
+            to: cal.date(from: DateComponents(year: 2026, month: 3, day: 7))!
+        )
+        let sessionId = all[0].0.id
+
+        let newStart = cal.date(from: DateComponents(year: 2026, month: 3, day: 6, hour: 9))!
+        let updated = try await service.editSession(id: sessionId, newStart: newStart, newStop: nil, newDuration: 5400)
+
+        #expect(updated.startTime == newStart)
+        #expect(abs(updated.endTime!.timeIntervalSince(newStart) - 5400) < 1)
+    }
+
+    @Test("edit with --stop + --duration sets stop and computes start")
+    func editStopAndDuration() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await insertSession(sessionRepo, projectId: project.id, startHour: 10, endHour: 12)
+
+        let all = try await sessionRepo.getSessions(
+            from: cal.date(from: DateComponents(year: 2026, month: 3, day: 6))!,
+            to: cal.date(from: DateComponents(year: 2026, month: 3, day: 7))!
+        )
+        let sessionId = all[0].0.id
+
+        let newStop = cal.date(from: DateComponents(year: 2026, month: 3, day: 6, hour: 14))!
+        let updated = try await service.editSession(id: sessionId, newStart: nil, newStop: newStop, newDuration: 7200)
+
+        #expect(abs(updated.startTime.timeIntervalSince(newStop) + 7200) < 1)
+        #expect(updated.endTime == newStop)
+    }
+
+    @Test("edit with all three flags throws overdetermined")
+    func editOverdetermined() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await insertSession(sessionRepo, projectId: project.id, startHour: 10, endHour: 12)
+
+        let all = try await sessionRepo.getSessions(
+            from: cal.date(from: DateComponents(year: 2026, month: 3, day: 6))!,
+            to: cal.date(from: DateComponents(year: 2026, month: 3, day: 7))!
+        )
+        let sessionId = all[0].0.id
+
+        let newStart = cal.date(from: DateComponents(year: 2026, month: 3, day: 6, hour: 9))!
+        let newStop = cal.date(from: DateComponents(year: 2026, month: 3, day: 6, hour: 11))!
+
+        await #expect(throws: RockyCoreError.self) {
+            try await service.editSession(id: sessionId, newStart: newStart, newStop: newStop, newDuration: 3600)
+        }
+    }
+
+    @Test("edit non-existent session throws sessionNotFound")
+    func editNotFound() async throws {
+        let (_, _, service) = makeServices()
+
+        await #expect(throws: RockyCoreError.self) {
+            try await service.editSession(id: 999, newStart: Date(), newStop: nil, newDuration: nil)
+        }
+    }
+
+    @Test("edit stop of running session throws cannotEditRunningSessionStop")
+    func editRunningStop() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await sessionRepo.start(projectId: project.id)
+
+        let running = try await sessionRepo.getRunning()
+        let sessionId = running[0].id
+
+        await #expect(throws: RockyCoreError.self) {
+            try await service.editSession(id: sessionId, newStart: nil, newStop: Date(), newDuration: nil)
+        }
+    }
+
+    @Test("edit start in future throws startTimeInFuture")
+    func editFutureStart() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await insertSession(sessionRepo, projectId: project.id, startHour: 10, endHour: 12)
+
+        let all = try await sessionRepo.getSessions(
+            from: cal.date(from: DateComponents(year: 2026, month: 3, day: 6))!,
+            to: cal.date(from: DateComponents(year: 2026, month: 3, day: 7))!
+        )
+        let sessionId = all[0].0.id
+
+        let futureStart = Date().addingTimeInterval(86400)
+
+        await #expect(throws: RockyCoreError.self) {
+            try await service.editSession(id: sessionId, newStart: futureStart, newStop: nil, newDuration: nil)
+        }
+    }
+
+    @Test("edit with stop before start throws stopBeforeStart")
+    func editStopBeforeStart() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await insertSession(sessionRepo, projectId: project.id, startHour: 10, endHour: 12)
+
+        let all = try await sessionRepo.getSessions(
+            from: cal.date(from: DateComponents(year: 2026, month: 3, day: 6))!,
+            to: cal.date(from: DateComponents(year: 2026, month: 3, day: 7))!
+        )
+        let sessionId = all[0].0.id
+
+        let badStop = cal.date(from: DateComponents(year: 2026, month: 3, day: 6, hour: 9))!
+
+        await #expect(throws: RockyCoreError.self) {
+            try await service.editSession(id: sessionId, newStart: nil, newStop: badStop, newDuration: nil)
+        }
+    }
+
+    @Test("edit with zero duration throws durationNotPositive")
+    func editZeroDuration() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await insertSession(sessionRepo, projectId: project.id, startHour: 10, endHour: 12)
+
+        let all = try await sessionRepo.getSessions(
+            from: cal.date(from: DateComponents(year: 2026, month: 3, day: 6))!,
+            to: cal.date(from: DateComponents(year: 2026, month: 3, day: 7))!
+        )
+        let sessionId = all[0].0.id
+
+        await #expect(throws: RockyCoreError.self) {
+            try await service.editSession(id: sessionId, newStart: nil, newStop: nil, newDuration: 0)
+        }
+    }
+
+    @Test("edit with negative duration throws durationNotPositive")
+    func editNegativeDuration() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await insertSession(sessionRepo, projectId: project.id, startHour: 10, endHour: 12)
+
+        let all = try await sessionRepo.getSessions(
+            from: cal.date(from: DateComponents(year: 2026, month: 3, day: 6))!,
+            to: cal.date(from: DateComponents(year: 2026, month: 3, day: 7))!
+        )
+        let sessionId = all[0].0.id
+
+        await #expect(throws: RockyCoreError.self) {
+            try await service.editSession(id: sessionId, newStart: nil, newStop: nil, newDuration: -100)
+        }
+    }
+
+    @Test("edit start of running session is allowed")
+    func editRunningStart() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await sessionRepo.start(projectId: project.id)
+
+        let running = try await sessionRepo.getRunning()
+        let sessionId = running[0].id
+
+        let newStart = Date().addingTimeInterval(-7200)
+        let updated = try await service.editSession(id: sessionId, newStart: newStart, newStop: nil, newDuration: nil)
+
+        #expect(abs(updated.startTime.timeIntervalSince(newStart)) < 1)
+        #expect(updated.isRunning)
+    }
+
+    @Test("edit session spanning midnight")
+    func editMidnightSession() async throws {
+        let (projectRepo, sessionRepo, service) = makeServices()
+        let project = try await projectRepo.findOrCreate(name: "test")
+        try await insertSession(sessionRepo, projectId: project.id, startHour: 23, startDay: 5, endHour: 10, endDay: 6)
+
+        let all = try await sessionRepo.getSessions(
+            from: cal.date(from: DateComponents(year: 2026, month: 3, day: 5))!,
+            to: cal.date(from: DateComponents(year: 2026, month: 3, day: 7))!
+        )
+        let sessionId = all[0].0.id
+
+        let newStop = cal.date(from: DateComponents(year: 2026, month: 3, day: 6, hour: 1, minute: 30))!
+        let updated = try await service.editSession(id: sessionId, newStart: nil, newStop: newStop, newDuration: nil)
+
+        #expect(updated.endTime == newStop)
+        #expect(updated.duration() == 9000) // 2.5 hours
+    }
+}
